actually move all functions related to running benchmarks to csbench_run.c (they were split between csbench_run.c and csbench.c)
