+++ csbench_report.c
@@ -54,11 +54,8 @@
 #include "csbench.h"
 
 #include <assert.h>
-#include <dirent.h>
 #include <errno.h>
-#include <fcntl.h>
 #include <signal.h>
-#include <stdarg.h>
 #include <stdlib.h>
 #include <sys/stat.h>
 #include <sys/wait.h>
